% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_funmediation.r
\name{print.funmediation}
\alias{print.funmediation}
\title{print.funmediation:  Print output from a model that was fit by the funmediation function.}
\usage{
\method{print}{funmediation}(x, ...)
}
\arguments{
\item{x}{The funmediation object (output of the funmediation function)}

\item{...}{Further arguments currently not supported}
}
\description{
print.funmediation:  Print output from a model that was fit by the funmediation function.
}
