% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_funmediation_example.R
\name{simulate_funmediation_example}
\alias{simulate_funmediation_example}
\title{simulate_funmediation_example function}
\usage{
simulate_funmediation_example(
  nsub = 500,
  ntimes = 100,
  observe_rate = 0.4,
  alpha_int = function(t) {     return(t^0.5) },
  alpha_X = function(t) {     return(-(t/2)^0.5) },
  beta_M = function(t) {     (1/2) * (exp(t) - 1) },
  beta_int = 0,
  beta_X = 0.2,
  sigma_Y = 1,
  sigma_M_error = 2,
  rho_M_error = 0.8,
  simulate_binary_Y = FALSE,
  make_covariate_S = FALSE
)
}
\arguments{
\item{nsub}{Number of subjects}

\item{ntimes}{Number of potential times that could be observed on each subject}

\item{observe_rate}{Proportion of potential times on which there are actually
observations. Not all times are observed; this is assumed to be completely random and
to be done by design to reduce participant burden.}

\item{alpha_int}{Function representing the time-varying mean of mediator variable
for the X=0 group}

\item{alpha_X}{Function representing the time-varying effect of X on the mediator}

\item{beta_M}{Function representing the functional coefficient for cumulative
(scalar-on-function) effect of M on Y adjusting for X}

\item{beta_int}{Mean of Y if X is zero and M is the 0 function}

\item{beta_X}{Direct effect of X on Y after adjusting for M}

\item{sigma_Y}{Error standard deviation of the outcome Y}

\item{sigma_M_error}{Error standard deviation of the mediator M}

\item{rho_M_error}{Autoregressive correlation coefficient of the mediator M
from one observation to the next}

\item{simulate_binary_Y}{Whether Y should be generated from a binary
logistic (TRUE) or Gaussian (FALSE) model}

\item{make_covariate_S}{Whether to generate a random binary covariate S
at the subject (i.e., time-invariant) level. It will be generated to have 
zero population-level relationship to the outcome.}
}
\value{
A list with the following components:
\describe{
\item{time_grid}{The time grid for interpreting functional coefficients.}
\item{true_alpha_int}{True value of the time-varying alpha_int parameter,
representing the time-specific mean of the mediator M when X is 0.}
\item{true_alpha_X}{True value of the time-varying alpha_X parameter,
representing the effect of X on M.}
\item{true_beta_int}{True value of the beta_M parameter, representing
the mean of the outcome, Y, when X=0 and M=0.}
\item{true_beta_M}{True value of the beta_M parameter, representing the
functional effect of the mediator, M, on the outcome, Y.}
\item{true_beta_X}{True value of the beta_X parameter, representing the
effect of X on the outcome, Y, adjusting for the mediator, M.}
\item{true_indirect}{True value of the indirect parameter, representing
the indirect (mediated) effect of X on the outcome, Y.}
\item{dataset}{The simulated longitudinal dataset in long form.}
}
}
\description{
Simulates a dataset for demonstrating the funmediation function.
}
